Refactor dollysearchservice to v1 and update OpenSearch integration #deploy-test-dolly-search-service #deploy-dolly-search-service
