+++ apps/dolly-search-service/config.test.yml
@@ -28,8 +28,6 @@ spec:
         - application: dolly-idporten
         - application: dolly-frontend
         - application: dolly-frontend-dev
-        - application: dolly-backend
-        - application: dolly-backend-dev
         - application: team-dolly-lokal-app
         - application: testnav-oversikt-frontend
     outbound:
