+++ apps/dolly-search-service/src/main/java/no/nav/testnav/dollysearchservice/dto/SearchRequest.java
@@ -5,7 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import no.nav.testnav.libs.data.dollysearchservice.v1.PersonRequest;
-import no.nav.testnav.libs.data.dollysearchservice.v2.ElasticTyper;
+import no.nav.testnav.libs.data.dollysearchservice.v1.ElasticTyper;
 
 import java.util.ArrayList;
 import java.util.HashSet;
