+++ apps/dolly-search-service/src/main/java/no/nav/testnav/dollysearchservice/mapper/OpenSearchV2ResponseMappingStrategy.java
@@ -4,7 +4,7 @@ import ma.glasnost.orika.CustomMapper;
 import ma.glasnost.orika.MapperFactory;
 import ma.glasnost.orika.MappingContext;
 import no.nav.testnav.dollysearchservice.dto.SearchInternalResponse;
-import no.nav.testnav.libs.data.dollysearchservice.v2.SearchResponse;
+import no.nav.testnav.libs.data.dollysearchservice.v1.SearchResponse;
 import org.springframework.stereotype.Component;
 
 @Component
