+++ apps/dolly-search-service/src/main/java/no/nav/testnav/dollysearchservice/provider/OpensearchController.java
@@ -4,13 +4,17 @@ import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import no.nav.testnav.dollysearchservice.service.OpenSearchService;
+import no.nav.testnav.libs.data.dollysearchservice.v1.ElasticTyper;
 import no.nav.testnav.libs.data.dollysearchservice.v1.SearchRequest;
 import no.nav.testnav.libs.data.dollysearchservice.v1.SearchResponse;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Mono;
+
+import java.util.List;
 
 @Slf4j
 @RestController
@@ -18,12 +22,13 @@ import reactor.core.publisher.Mono;
 @RequiredArgsConstructor
 public class OpensearchController {
 
-    private final OpenSearchService openSearchService;
+    private final OpenSearchService openSearchV2Service;
 
     @PostMapping
     @Operation(description = "Henter personer som matcher søk av persondetaljer i request")
-    public Mono<SearchResponse> getPersoner(@RequestBody SearchRequest request) {
+    public Mono<SearchResponse> getPersoner(@RequestParam(required = false) List<ElasticTyper> registreRequest,
+                                            @RequestBody SearchRequest request) {
 
-        return openSearchService.search(request);
+        return openSearchV2Service.search(request, registreRequest);
     }
 }
