+++ apps/dolly-search-service/src/main/java/no/nav/testnav/dollysearchservice/service/OpenSearchService.java
@@ -3,28 +3,39 @@ package no.nav.testnav.dollysearchservice.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import ma.glasnost.orika.MapperFacade;
-import no.nav.testnav.dollysearchservice.consumer.OpenSearchConsumer;
 import no.nav.testnav.dollysearchservice.utils.OpenSearchQueryBuilder;
+import no.nav.testnav.libs.data.dollysearchservice.v1.ElasticTyper;
 import no.nav.testnav.libs.data.dollysearchservice.v1.SearchRequest;
 import no.nav.testnav.libs.data.dollysearchservice.v1.SearchResponse;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
+
+import java.util.List;
+import java.util.Set;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class OpenSearchService {
 
-    private final OpenSearchConsumer openSearchConsumer;
+    private static final String NO_IDENT = "9999999999)";
+
+    private final BestillingQueryService bestillingQueryService;
     private final MapperFacade mapperFacade;
     private final PersonQueryService personQueryService;
 
-    public Mono<SearchResponse> search(SearchRequest searchRequest) {
+    public Mono<SearchResponse> search(SearchRequest searchRequest, List<ElasticTyper> registreRequest) {
 
-        var request = mapperFacade.map(searchRequest, no.nav.testnav.dollysearchservice.dto.SearchRequest.class);
+        var request = mapperFacade.map(searchRequest,
+                no.nav.testnav.dollysearchservice.dto.SearchRequest.class);
+        request.setRegistreRequest(registreRequest);
+
+        var identer = bestillingQueryService.execRegisterQuery(request);
+        request.setIdenter(identer.isEmpty() ? Set.of(NO_IDENT) : identer);
+
         var query = OpenSearchQueryBuilder.buildSearchQuery(request);
 
-        return personQueryService.execQuery(request, query)
+        return personQueryService.execQuery(request,query)
                 .map(response -> mapperFacade.map(response, SearchResponse.class));
     }
 }
